refactor(agents): split tool.ts and agent-loop.ts out of dynamic-agents
